end of day commit

+++ ft_strlcat.c
@@ -6,14 +6,16 @@
 /*   By: jvalkama <jvalkama@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/15 10:47:26 by jvalkama          #+#    #+#             */
-/*   Updated: 2025/04/15 11:02:08 by jvalkama         ###   ########.fr       */
+/*   Updated: 2025/04/17 09:55:49 by jvalkama         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-unsigned int	ft_strlcat(char *dest, char *src, unsigned int size)
+#include "libft.h"
+
+size_t	ft_strlcat(char *dest, char *src, size_t size)
 {
-	unsigned int	i;
-	unsigned int	j;
+	size_t	 i;
+	size_t	 j;
 
 	i = 0;
 	j = 0;
